Start work on a separate graph abstraction
This is part of an approach to make reading and writing similar in that they
work on the same sort of tree structure; We want to keep the EDM file data
as clean as possible and as separate from blender as we can.

For now, reading just reads the tree structure and uses it to print the tree.